# docs/integrations/aily-connector-openapi.yaml
openapi: 3.1.0
info:
  title: Quanxin Life Aily Connector
  version: phase1-v1
  description: Reference-only facade over audited Agent runs and ToolResults.
servers:
  - url: https://YOUR_REVIEWED_HOST
security:
  - connectorBearer: []
paths:
  /v1/aily/analysis-tasks:
    post:
      operationId: createAnalysisTask
      summary: Create an analysis task from a registered data batch
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAnalysisTaskRequest'
      responses:
        '202':
          description: Durable Agent run reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRunState'
        '401':
          description: Connector authentication failed
        '422':
          description: Task request rejected
  /v1/aily/analysis-tasks/{run_id}:
    get:
      operationId: getAnalysisTask
      summary: Get the durable state of an analysis task
      parameters:
        - $ref: '#/components/parameters/RunId'
      responses:
        '200':
          description: Existing Agent run state
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentRunState'
        '401':
          description: Connector authentication failed
        '404':
          description: Task not found
  /v1/aily/analysis-tasks/{run_id}/results/{result_id}:
    get:
      operationId: getAuditedResult
      summary: Get a run-bound ledger-registered ToolResult
      parameters:
        - $ref: '#/components/parameters/RunId'
        - name: result_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Revalidated ToolResult including exact provenance and versions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ToolResult'
        '401':
          description: Connector authentication failed
        '404':
          description: Result is absent, invalid, or not bound to the run
        '409':
          description: Result exists but its model route or audit authorization is not active
  /v1/aily/analysis-tasks/{run_id}/reports/{report_result_id}:
    get:
      operationId: getAuditedReport
      summary: Download a run-bound audited Markdown report
      parameters:
        - $ref: '#/components/parameters/RunId'
        - name: report_result_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: SHA-verified audited report artifact
          content:
            text/markdown:
              schema:
                type: string
        '401':
          description: Connector authentication failed
        '404':
          description: Report is absent, invalid, or not bound to the run
        '409':
          description: Report result is not authorized or its artifact integrity check failed
components:
  securitySchemes:
    connectorBearer:
      type: http
      scheme: bearer
  parameters:
    RunId:
      name: run_id
      in: path
      required: true
      schema:
        type: string
  schemas:
    CreateAnalysisTaskRequest:
      type: object
      additionalProperties: false
      required: [task_type, data_batch_id]
      properties:
        task_type:
          type: string
          enum:
            - predict_cycle_life
            - predict_soh_trajectory
            - compare_operation_scenarios
            - ingest_observed_soh
            - update_trajectory
        data_batch_id:
          type: string
    AgentRunState:
      type: object
      description: Existing quanxin_life.core.AgentRunState contract.
      required: [run_id, intent_id, plan_hash, status, result_ids, updated_at]
      additionalProperties: true
      properties:
        run_id: {type: string}
        intent_id: {type: string}
        plan_hash: {type: string}
        status: {type: string}
        result_ids:
          type: array
          items: {type: string}
        updated_at: {type: string, format: date-time}
    ToolResult:
      type: object
      description: Existing quanxin_life.core.ToolResult contract resolved from the audit ledger.
      required:
        - result_id
        - tool_name
        - tool_version
        - input_hash
        - values
        - warnings
        - provenance
        - created_at
      additionalProperties: true
      properties:
        result_id: {type: string}
        tool_name: {type: string}
        tool_version: {type: string}
        model_version: {type: [string, 'null']}
        data_version: {type: [string, 'null']}
        feature_version: {type: [string, 'null']}
        input_hash: {type: string}
        values: {type: object}
        warnings:
          type: array
          items: {type: string}
        provenance:
          type: array
          items: {type: object}
        created_at: {type: string, format: date-time}